+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Parse raw values to values for all fields (#98).
 - Add constraints and tests for Direction (#97).
 - Improve support for Distance, Direction and Elapsed Time (#96).
 - Refactor exceptions (#94).